fix for python 3

+++ naive_bayes.py
@@ -26,7 +26,7 @@ class Classifier(object):
         probability = Decimal('1')
         for token in tokens:
             counts = Decimal(token_counts.get(token, 0) + delta)
-            token_probability = counts / (total_words_in_category + (delta * vocab_size))
+            token_probability = counts / Decimal(total_words_in_category + (delta * vocab_size))
             probability = probability * token_probability
 
         return probability
